web/ssti_flask/requirements.txt
flask
gunicorn
mysql-connector-python